refactor(browser-agent): extract testpages + wsframe from package main
The /tests/ self-test fixture server (embedded pages, HTTP error fixtures,
WebSocket echo harness) had exactly one entry point into main —
setupHTTPRoutes registering handleTestPages() and handleTestHarnessWS — so it
moves out whole as cmd/browser-agent/internal/testpages.

*** The filename lied about one file. ***

testpages_websocket_codec.go was named as part of the self-testing feature, but
the reference graph says otherwise: terminal_bridge.go injects wsReadFrame,
wsWriteFrame and wsAcceptKey into terminal.Deps, so the production terminal
relay depends on that codec. It is shared RFC 6455 wire infrastructure, not a
test fixture, and it is now cmd/browser-agent/internal/wsframe with both
consumers importing it. Grouping by prefix would have buried a production
dependency inside a package called "testpages".

Two more prefix/graph disagreements resolved while rewiring:
  - jsonResponse (main, handler_http.go) -> util.JSONResponse, the canonical
    helper internal/server already delegates to. Same headers, same status,
    same encoder; main's copy differs only in routing encode errors through the
    swappable stderrSink, which no /tests/ response path can reach.
  - buildQueryParams (main, tools_response.go — an MCP *tool response* helper)
    was building the WebSocket echo envelope. Replaced with the mcp.SafeMarshal
    call it forwards to, dropping a tools->testpages edge that never made sense.

Tests moved with their code and were split along the same seam: the frame
codec tests are now wsframe/frame_test.go, the HTTP/handshake tests stay in
testpages. No test references the old symbols.

cmd/browser-agent: 191 -> 188 non-test files.
