+++ cmd/browser-agent/internal/testpages/pages/cdp-smoke-test.html
