+++ cmd/browser-agent/internal/testpages/pages/interaction-test.html
