+++ cmd/browser-agent/internal/testpages/pages/performance.html
