+++ cmd/browser-agent/internal/testpages/pages/recording.html
