+++ cmd/browser-agent/internal/testpages/pages/telemetry.html
